backend/requirements.txt
fastapi
uvicorn[standard]
PyMySQL
SQLAlchemy
python-dotenv
passlib[bcrypt]
bcrypt
pydantic[email]
python-jose[pyjwt]
pydantic-settings
python-multipart
motor 
cryptography 
pymongo